Commit pasteFromExcel rmd and md
Used RStudio to create the blog post I wanted. Copied to StackEdit. In StackEdit, replaced url's to images as necessary, inserted console output where appropriate because, as an md file r.t. an rmd file, R wasn't going to be available to generate that on the fly. In StackEdit, published the html to Tirnostics' blog "Page" and then copied that html into a new bolog "Post".

Don't know why I needed a "Page" first, but at least I got it to work and I can play around with going directly to a "Post" next time.